Moved downloading svg code to separate file
